Add a logger to modules by default.
Each module now has a $_logger attribute which is a Plop_Logger instance
and can be used to log useful stuff. See Plop's documentation for more
information.
Also, this logger is only defined when Plop is available and is set to
NULL otherwise. The recommended usage pattern in the code is:
```
$this->_logger and $this->_logger->debug("Some debugging message");
```

This way, your code won't break whenever Plop hasn't been installed.