Remove getTranslator($chan) from Erebot_Module_Base.
It's been replaced with getFormatter($chan) which returns a formatter
(Erebot_Interface_Styling) instead. The formatter is built using the
same translator as what getTranslator($chan) used to return.
Also, it has a getTranslator() method.

So ``$this->getTranslator($chan);`` can effectively be replaced with:
``$this->getFormatter($chan)->getTranslator();`` in case you really
need a translator (which most of the time you don't).

This change also reduces the overhead of calling a formatter's _()
method instead of the translator's gettext() method for really basic
strings (where no markup appears).

Last but not least, this change adds some more doc to the API.