Change _unload() to a dummy implementation.
Most of the time, _unload() won't be redefined by subclasses, so it's plain
stupid to force developpers to implement it (using the abstract keyword).

Instead, we provide a dummy implementation that does nothing and leave the
possibility for subclasses to override this method to add their own logic
if they need to.