Rethink profiles for numeric messages.
Completely rethink the way we associate irc servers with the numerics they
support. We now define a specific profile for major ircds (with the specific
additional numerics they may define) and a generic profile that tries to
follow RFC 2812 for the other ircds out there.

The following IRCds currently have their own profile:
- bahamut
- csircd
- InspIRCd
- irc2
- ircd-hybrid
- ircu
- UltimateIRCd
- UnrealIRCd
These profiles are part of the core of Erebot. See the code there for more
information.